Replace words with Digits
We have verified we need to tackle overlapping words. Tests with overlapping words are failing.

Go over the string and assign index values to each number to represent the first instance of each in the string. Then we'll replace the word with the number that has the smallest index value. Repeats until done

- [x] character sequence
- [x] exception flow control
- [x] list literal
- [x] multiline
- [x] Only read first and last digits
- [x] Refactor for loops
- [x] Optionals
- [x] widen
- [ ] move line
- [ ] kill line
- [ ] camelhumps
- [x] test pyramid
- [x] immutable list
- [ ] concurrent modification
- [x] final parameters
- [ ] change test part two to express how the result is computed
- [ ] consider test coverage

Co-authored-by: Mark Crossfield <mark.crossfield@autotrader.co.uk>