Starting Day01 2022
Starting off working on day01 part1 of Advent of Code year 2022. Trying to write code in a functional way. Thinking of maybe splitting by comma's to create a list containing a list of integers. Then reducing the inner lists before joining them back. Which we can then return the Max to solve part1

- [x] character sequence
- [x] exception flow control
- [x] list literal
- [x] multiline
- [x] Only read first and last digits
- [x] Refactor for loops
- [x] Optionals
- [x] widen
- [x] move line
- [x] kill line
- [x] camelhumps
- [x] test pyramid
- [x] immutable list
- [ ] concurrent modification
- [x] final parameters
- [x] change test part two to express how the result is computed
- [ ] consider test coverage

Co-authored-by: Mark Crossfield <mark.crossfield@autotrader.co.uk>